# internal/delivery/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/manager/get_case": {
            "get": {
                "description": "Retrieves a case by its ID and returns detailed information about the case.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "managers"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of the case to retrieve",
                        "name": "case_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the case details",
                        "schema": {
                            "$ref": "#/definitions/models.CaseFul"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameter or missing case_id",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "403": {
                        "description": "JWT is invalid or expired",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    }
                }
            }
        },
        "/manager/get_specialists_rating": {
            "get": {
                "description": "Retrieves a list of specialists' ratings within a specified time range, paginated by a cursor.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "managers"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Cursor for pagination",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time for filtering ratings (inclusive), in RFC3339 format",
                        "name": "time_from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End time for filtering ratings (inclusive), in RFC3339 format",
                        "name": "time_to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the specialists' ratings",
                        "schema": {
                            "$ref": "#/definitions/models.RatingSpecialistCountCursor"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameter or missing required fields",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "403": {
                        "description": "JWT is invalid or expired",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CaseFul": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "datetime": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "photo_url": {
                    "type": "string"
                },
                "rated_covers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RatedCover"
                    }
                },
                "transport": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "violation_value": {
                    "type": "string"
                }
            }
        },
...
        "models.RatedCover": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "fullname": {
                    "$ref": "#/definitions/null.String"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "photo_url": {
                    "$ref": "#/definitions/null.String"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.RatingSpecialistCount": {
            "type": "object",
            "properties": {
                "correct": {
                    "type": "integer"
                },
                "fullname": {
                    "$ref": "#/definitions/null.String"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "photo_url": {
                    "$ref": "#/definitions/null.String"
                },
                "total": {
                    "type": "integer"
                },
                "unknown": {
                    "type": "integer"
                }
            }
        },
        "models.RatingSpecialistCountCursor": {
            "type": "object",
            "properties": {
                "cursor": {
                    "$ref": "#/definitions/null.Int"
                },
                "specialists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RatingSpecialistCount"
                    }
                }
            }
        },
...
        "null.Int": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
        "null.String": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "responses.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}